Revert "vfs_fruit: fix fruit:resource option spelling, but not behaviour"
This reverts commit 586f8b73b6c189cb825ce2e8817230ca57f87d35.

This patchset slipped in by accident and is not in master yet.
The commit message does not fit to the set of patches, because numerous patches
have been squashed into one commit. It's a rewrite of vfs_fruit.
Reverting until review will be finished and patches are in master.

Autobuild-User(v4-5-test): Karolin Seeger <kseeger@samba.org>
Autobuild-Date(v4-5-test): Fri Jan 27 09:48:52 CET 2017 on sn-devel-144